fix: 장바구니 품목 추가,수량 증가 시 dto의 id type string으로 수정

+++ src/main/kotlin/com/liveforpresent/cookiosk/api/cart/command/presentation/CartCommandController.kt
@@ -66,7 +66,7 @@ class CartCommandController(
             price = reqDto.price,
             imageUrl = reqDto.imageUrl,
             quantity = reqDto.quantity,
-            productId = reqDto.productId
+            productId = reqDto.productId.toLong()
         )
 
         addItemHandler.execute(command)
